clean(ZMSKVR-436): organize test files to match src folder structure
- Move Appointment component tests to tests/unit/Appointment/
- Move API tests to tests/unit/api/
- Move utility tests to tests/unit/utils/
- Maintain existing AppointmentDetail/ and AppointmentOverview/ folders
- Improve test organization and maintainability
